refactor(tools/xnvmeperf): move to subdir and extract shared header
Move xnvmeperf.c into tools/xnvmeperf/ and introduce xnvmeperf.h with
the shared types (enum iopattern, struct xnvmeperf_args) and
fill_pattern, in preparation for adding a GPU path that shares
these definitions. Extract the xnvmeperf meson build into its own
meson.build and replace the inline block in tools/meson.build with
subdir('xnvmeperf').
