+++ tools/meson.build
@@ -59,17 +59,4 @@ foreach source, tests : tools
   endforeach
 endforeach
 
-if thread_dep.found()
-  executable(
-    'xnvmeperf',
-    'xnvmeperf.c',
-    include_directories: [conf_inc, xnvme_inc],
-    link_args: link_args_hardening,
-    link_with: xnvmelib,
-    install_rpath: xnvmelib_rpath,
-    dependencies: [thread_dep],
-    install: true,
-  )
-else
-  warning('No support for threads, skipping xnvmeperf')
-endif
+subdir('xnvmeperf')
